% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bookmark_to_channel.R
\name{add_bookmark_to_channel}
\alias{add_bookmark_to_channel}
\title{add_bookmark_to_channel}
\usage{
add_bookmark_to_channel(
  channel,
  bookmark,
  all_channel = slackr::slackr_channels(),
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{channel}

\item{bookmark}{as list}

\item{token}{slack api token}
}
\description{
add_bookmark_to_channel
}
\examples{


\dontrun{

add_bookmark_to_channel(
channel = "bac_a_sable_2010bis",
bookmark = list(
  title = "hophophop4",
  link = "https://www.google.fr",
  emoji = ":kekette:"
)
)






}






}
