% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{channel_exist}
\alias{channel_exist}
\title{doest this channel exist ?}
\usage{
channel_exist(name, channel = slackr::slackr_channels())
}
\arguments{
\item{name}{name}

\item{channel}{channel}
}
\description{
doest this channel exist ?
}
