% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_channel_id}
\alias{get_channel_id}
\title{get_channel_id}
\usage{
get_channel_id(name, channel = slackr::slackr_channels())
}
\arguments{
\item{name}{name}

\item{channel}{channel}
}
\description{
get_channel_id
}
