% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_user_to_channel.R
\name{invite_single_user_to_channel}
\alias{invite_single_user_to_channel}
\title{Add a single user to a channel}
\usage{
invite_single_user_to_channel(
  channel,
  user,
  token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)
}
\arguments{
\item{channel}{channel name}

\item{user}{user name}

\item{token}{slack api token}
}
\description{
Add a single user to a channel
}
