% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_user_to_channel.R
\name{invite_user_to_channel}
\alias{invite_user_to_channel}
\title{Title}
\usage{
invite_user_to_channel(user, channel,
  token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{token}{}
}
\description{
Title
}
