% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_user_to_channel.R
\name{invite_users_to_channel}
\alias{invite_users_to_channel}
\title{Add user(s) to a channel}
\usage{
invite_users_to_channel(channel, users, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{channel}{channel name}

\item{users}{vector of user's name}

\item{token}{slack api token}
}
\description{
Add user(s) to a channel
}
