% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_channel_topic.R
\name{set_channel_topic}
\alias{set_channel_topic}
\title{Title}
\usage{
set_channel_topic(channel, topic, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{token}{}
}
\description{
Title
}
