% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bookmark_to_channel.R
\name{add_bookmarks_to_channel}
\alias{add_bookmarks_to_channel}
\title{Title}
\usage{
add_bookmarks_to_channel(
  channel,
  bookmarks,
  all_channel = slackr::slackr_channels(),
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{channel}

\item{bookmarks}{bookmarks}

\item{token}{token}
}
\description{
Title
}
