% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slack_channel.R
\name{create_slack_channel}
\alias{create_slack_channel}
\title{Title}
\usage{
create_slack_channel(channel, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{token}{}
}
\description{
Title
}
