% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slack_channel.R
\name{create_slack_channel}
\alias{create_slack_channel}
\title{Create a new slack channel}
\usage{
create_slack_channel(channel, token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
}
\arguments{
\item{channel}{channel name}

\item{token}{slack API token}
}
\value{
channel name as character
}
\description{
Create a new slack channel
}
