% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bookmark_starting_with.R
\name{remove_bookmark_starting_with}
\alias{remove_bookmark_starting_with}
\title{Title}
\usage{
remove_bookmark_starting_with(
  pattern,
  channel,
  all = slack::get_all_bookmarks_from_channel(channel = channel, token = token,
    all_channel = all_channel),
  all_channel = slackr::slackr_channels(),
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{token}{}
}
\description{
Title
}
